[Numerics] Create intermediate class for sparse Jacobians
