+++ include/cantera/numerics/AdaptivePreconditioner.h
@@ -10,10 +10,7 @@
 #ifndef ADAPTIVEPRECONDITIONER_H
 #define ADAPTIVEPRECONDITIONER_H
 
-#include "cantera/numerics/SystemJacobian.h"
-#include "cantera/numerics/eigen_sparse.h"
-#include "cantera/base/global.h"
-#include <iostream>
+#include "cantera/numerics/EigenSparseJacobian.h"
 
 namespace Cantera
 {
@@ -23,7 +20,7 @@ namespace Cantera
 //! the preconditioner by a threshold value. It also neglects pressure
 //! dependence and third body contributions in its formation and has a
 //! finite difference approximation for temperature.
-class AdaptivePreconditioner : public SystemJacobian
+class AdaptivePreconditioner : public EigenSparseJacobian
 {
 public:
     AdaptivePreconditioner();
@@ -31,7 +28,7 @@ public:
     void initialize(size_t networkSize) override;
 
     void reset() override {
-        m_precon_matrix.setZero();
+        m_matrix.setZero();
         m_jac_trips.clear();
     };
 
@@ -39,10 +36,7 @@ public:
     void setup() override; // deprecated
     void factorize() override;
     void solve(const size_t stateSize, double* rhs_vector, double* output) override;
-    void setValue(size_t row, size_t col, double value) override;
     void stateAdjustment(vector<double>& state) override;
-    void updatePreconditioner() override;
-    void updateTransient(double rdt, int* mask) override;
 
     int info() const override {
         return static_cast<int>(m_solver.info());
@@ -50,20 +44,6 @@ public:
 
     //! Prune preconditioner elements
     void prunePreconditioner();
-
-    //! Return semi-analytical Jacobian of an AdaptivePreconditioner object.
-    //! @ingroup derivGroup
-    Eigen::SparseMatrix<double> jacobian() {
-        Eigen::SparseMatrix<double> jacobian_mat(m_dim, m_dim);
-        jacobian_mat.setFromTriplets(m_jac_trips.begin(), m_jac_trips.end());
-        return jacobian_mat;
-    }
-
-    //! Return the internal preconditioner matrix
-    Eigen::SparseMatrix<double> matrix() {
-        updatePreconditioner();
-        return m_precon_matrix;
-    }
 
     //! Get the threshold value for setting elements
     double threshold() { return m_threshold; }
@@ -95,27 +75,12 @@ public:
         m_solver.setFillfactor(fillFactor);
     }
 
-    //! Print preconditioner contents
-    void printPreconditioner() override;
-
-    //! Print jacobian contents
-    void printJacobian();
-
 protected:
     //! ILUT fill factor
     double m_fill_factor = 0;
 
     //! ILUT drop tolerance
     double m_drop_tol = 0;
-
-    //! Vector of triples representing the jacobian used in preconditioning
-    vector<Eigen::Triplet<double>> m_jac_trips;
-
-    //! Storage of appropriately sized identity matrix for making the preconditioner
-    Eigen::SparseMatrix<double> m_identity;
-
-    //! Container that is the sparse preconditioner
-    Eigen::SparseMatrix<double> m_precon_matrix;
 
     //! Solver used in solving the linear system
     Eigen::IncompleteLUT<double> m_solver;
